Add user management feature
- Add a new admin site (/admin/) which is used to register users for this site.
- The unregistered users are redirected to the no_authority page.

+++ timestamp.py
@@ -10,27 +10,28 @@ from google.appengine.ext import ndb
 
 import jinja2
 import webapp2
+import logging
+
+from data_modules import Timestamp
+from data_modules import User
 
 JINJA_ENVIRONMENT = jinja2.Environment(
 	loader = jinja2.FileSystemLoader(os.path.dirname(__file__)),
 	extensions = ['jinja2.ext.autoescape'],
 	autoescape = True)
 
-# We set a parent key on the 'Username' to ensure that they are all in the same
+# We set a parent key on the 'User' to ensure that they are all in the same
 # entity group. Queries across the single entity group will be consistent.
-# However, the write rate should be limited to ~1/second.
-
 def user_key(user):
-	return ndb.Key('Username', user.user_id())
-	
-	
-class Timestamp(ndb.Model):
-	"""Models an individual Timestamp entry."""
-	start = ndb.DateTimeProperty(auto_now_add=True)
-	finish = ndb.DateTimeProperty(auto_now_add=False)
-	user = ndb.UserProperty(indexed=False)
-	content = ndb.StringProperty(indexed=False)
+	return ndb.Key('User', user.user_id())
 
+def checkAuthority(self):
+	nickname = users.get_current_user().nickname()
+	user_key = ndb.Key('User', nickname)
+	users_query = User(key = user_key).query()
+	user = users_query.fetch(1)[0]
+	if user.nickname != nickname:
+		self.redirect("/no_authority")
 	
 def getTimestamps():
 	user = users.get_current_user()
@@ -40,7 +41,8 @@ def getTimestamps():
 
 	
 class MainPage(webapp2.RequestHandler):
-    def get(self):
+	def get(self):
+		checkAuthority(self)
 		timestamps = getTimestamps()
 		utc = pytz.timezone('UTC')
 		kst = pytz.timezone('Asia/Seoul')
@@ -59,6 +61,7 @@ class MainPage(webapp2.RequestHandler):
 
 class Checkin(webapp2.RequestHandler):
 	def post(self):
+		checkAuthority(self)
 		user = users.get_current_user();
 		timestamp = Timestamp(parent=user_key(user))
 		timestamp.put()
@@ -67,6 +70,7 @@ class Checkin(webapp2.RequestHandler):
 		
 class Checkout(webapp2.RequestHandler):
 	def post(self):
+		checkAuthority(self)
 		timestamps = getTimestamps()
 		if len(timestamps) > 0:
 			timestamp = timestamps[0]
@@ -79,16 +83,30 @@ class Checkout(webapp2.RequestHandler):
 		
 class Cancel(webapp2.RequestHandler):
 	def post(self):
+		checkAuthority(self)
 		timestamps = getTimestamps()
 		if len(timestamps) > 0:
 			timestamp = timestamps[0]
 			timestamp.key.delete()
 		self.redirect('/')
 		
+class NoAuthority(webapp2.RequestHandler):
+	def get(self):
+		logging.info(self.request.uri)
+		logout_url = users.create_logout_url('/')
+		
+		template_values = {
+			'logout_url': logout_url,
+		}
+		
+		template = JINJA_ENVIRONMENT.get_template('no_authority.html')
+		self.response.write(template.render(template_values))
+				
 
 application = webapp2.WSGIApplication([
     ('/', MainPage),
 	('/checkin', Checkin),
 	('/checkout', Checkout),
 	('/cancel', Cancel),
+	('/no_authority', NoAuthority),
 ], debug=True)
